+++ packages/ui/src/components/organisms/Form/DynamicForm/fields/DocumentField/index.ts
@@ -1,2 +1,3 @@
 export * from './DocumentField';
 export * from './helpers/is-document-field-definition';
+export * from './hooks/useDocumentUpload';