+++ packages/ui/src/components/organisms/Form/DynamicForm/layouts/FieldPriorityReason/FieldPriorityReason.tsx
@@ -1,4 +1,5 @@
 import { createTestId } from '@/components/organisms/Renderer';
+import { InfoIcon } from 'lucide-react';
 import { useStack } from '../../fields/FieldList/providers/StackProvider';
 import { usePriorityFields } from '../../hooks/internal/usePriorityFields';
 import { IFormElement } from '../../types';
@@ -11,14 +12,19 @@ export const FieldPriorityReason: React.FC<IFieldPriorityReasonProps> = ({ eleme
   const { stack } = useStack();
   const { priorityField } = usePriorityFields(element);
 
-  if (!priorityField) return null;
+  if (!priorityField) {
+    return null;
+  }
 
   return (
-    <p
-      className="py-1 text-[0.8rem] text-amber-400"
-      data-testid={`${createTestId(element, stack)}-priority-reason`}
-    >
-      {priorityField.reason}
-    </p>
+    <div className="flex flex-row flex-nowrap items-start gap-2">
+      <InfoIcon size={16} className="text-warning mt-1.5 shrink-0" />
+      <p
+        className="text-warning py-1 text-[0.8rem]"
+        data-testid={`${createTestId(element, stack)}-priority-reason`}
+      >
+        {priorityField.reason}
+      </p>
+    </div>
   );
 };