+++ packages/ui/src/components/organisms/Form/Validator/validators/index.ts
@@ -1,4 +1,5 @@
 import { TBaseValidators, TValidator } from '../types';
+import { documentValidator } from './document';
 import { formatValidator } from './format';
 import { maxLengthValidator } from './max-length';
 import { maximumValueValidator } from './maximum';
@@ -15,6 +16,7 @@ export const baseValidatorsMap: Record<TBaseValidators, TValidator<any, any>> = 
   minimum: minimumValueValidator,
   maximum: maximumValueValidator,
   format: formatValidator,
+  document: documentValidator,
 };
 
 export const validatorsExtends: Record<string, TValidator<any, any>> = {};