+++ apps/kyb-app/CHANGELOG.md
@@ -1,5 +1,14 @@
 # kyb-app
 
+## 0.3.139
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.79
+  - @ballerine/ui@0.7.110
+  - @ballerine/workflow-browser-sdk@0.6.100
+
 ## 0.3.138
 
 ### Patch Changes